updating for record going into game and moving functions out of main.py
